% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classifier_rpart}
\alias{classifier_rpart}
\alias{predicter_rpart}
\alias{classifier_glm}
\alias{predicter_glm}
\alias{classifier_nb}
\alias{predicter_nb}
\alias{classifier_earth}
\alias{predicter_earth}
\title{Functions to be used internally}
\usage{
classifier_rpart(x_train, y_train, weights, ...)

predicter_rpart(model, x_new, ...)

classifier_glm(x_train, y_train, weights, ...)

predicter_glm(model, x_new, ...)

classifier_nb(x_train, y_train, weights, ...)

predicter_nb(model, x_new, ...)

classifier_earth(x_train, y_train, weights, ...)

predicter_earth(model, x_new, ...)
}
\arguments{
\item{x_train}{input features.}

\item{y_train}{factor class variable.}

\item{weights}{instance weights.}

\item{...}{other control parameters.}

\item{model}{model obtained from respective classifier.}

\item{x_new}{new features for prediction.}
}
\value{
Classifiers produce an object which is appropriate
for respective predicter. Predicters returns class
predictions for \code{x_new}.
}
\description{
for internal use
}
\keyword{internal}
