% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_weighted.R
\name{mean_weighted}
\alias{mean_weighted}
\title{Weighted mean}
\usage{
mean_weighted(x, w = NULL)
}
\arguments{
\item{x}{a numerical vector.}

\item{w}{vector of weights. If NULL, non-weighted mean is calculated.}
}
\description{
Calculates weighted mean
}
\keyword{internal}
