% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.booster.R
\name{plot.booster}
\alias{plot.booster}
\alias{plot.discrete_adaboost}
\alias{plot.real_adaboost}
\title{plot booster}
\usage{
\method{plot}{booster}(x, y, ...)

\method{plot}{discrete_adaboost}(x, y, ...)

\method{plot}{real_adaboost}(x, y, ...)
}
\arguments{
\item{x}{booster object}

\item{y}{ignored}

\item{...}{additional arguments.}
}
\value{
Summary of "booster" object.
}
\description{
plots errors of booster model iterations
}
\keyword{internal}
