% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.booster.R
\name{predict.booster}
\alias{predict.booster}
\title{Prediction function for Adaboost framework}
\usage{
\method{predict}{booster}(object, newdata, type = "pred", print_detail = FALSE, ...)
}
\arguments{
\item{object}{booster object}

\item{newdata}{a factor class variable. Boosting algorithm allows for}

\item{type}{pre-ready or a custom classifier function.}
}
\value{
A vector of class predictions or a matrix of class probabilities
depending of \code{type}
}
\description{
Makes predictions based on booster function
}
\details{
Discrete Adaboost does not estimate class probabilities but here, "prob"
gives class probabilities. Probabilities are calculated using sum
of weighted class votes, that is votes/sum(votes). It has no theoretical
background. Use it on your own risk.
}
\keyword{internal}
