% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.booster.R
\name{print.booster}
\alias{print.booster}
\title{Print booster}
\usage{
\method{print}{booster}(x, ...)
}
\arguments{
\item{x}{booster object}
}
\value{
Summary of "booster" object.
}
\description{
Prints a summary of booster model
}
\keyword{internal}
