% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.w_naive_bayes.R
\name{print.w_naive_bayes}
\alias{print.w_naive_bayes}
\alias{print.w_gaussian_naive_bayes}
\alias{print.w_discrete_naive_bayes}
\title{Print w_naive_bayes}
\usage{
\method{print}{w_naive_bayes}(x, ...)

\method{print}{w_gaussian_naive_bayes}(x, ...)

\method{print}{w_discrete_naive_bayes}(x, ...)
}
\arguments{
\item{x}{w_naive_bayes object}

\item{...}{additional arguments.}
}
\value{
Summary of "w_naive_bayes" object.
}
\description{
Prints a summary of w_naive_bayes model
}
\keyword{internal}
