% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_weighted.R
\name{var_weighted}
\alias{var_weighted}
\title{Weighted variance}
\usage{
var_weighted(x, w = NULL, mean = NULL)
}
\arguments{
\item{x}{a numerical vector.}

\item{w}{vector of weights. If NULL, non-weighted variance is calculated.}

\item{mean}{a pre-calculated mean. If NULL, mean is calculated.}
}
\description{
Calculates weighted variance
}
\keyword{internal}
