% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classifier_rpart}
\alias{classifier_rpart}
\alias{predictor_rpart}
\alias{classifier_glm}
\alias{predictor_glm}
\alias{classifier_gnb}
\alias{predictor_gnb}
\alias{classifier_dnb}
\alias{predictor_dnb}
\alias{classifier_earth}
\alias{predictor_earth}
\title{Functions to be used internally}
\usage{
classifier_rpart(x_train, y_train, weights, ...)

predictor_rpart(model, x_new, type = "pred", ...)

classifier_glm(x_train, y_train, weights, ...)

predictor_glm(model, x_new, type = "pred", ...)

classifier_gnb(x_train, y_train, weights, ...)

predictor_gnb(model, x_new, type = "pred", ...)

classifier_dnb(x_train, y_train, weights, ...)

predictor_dnb(model, x_new, type = "pred", ...)

classifier_earth(x_train, y_train, weights, ...)

predictor_earth(model, x_new, type = "pred", ...)
}
\arguments{
\item{x_train}{input features.}

\item{y_train}{factor class variable.}

\item{weights}{instance weights.}

\item{...}{other control parameters.}

\item{model}{model obtained from respective classifier.}

\item{x_new}{new features for prediction.}
}
\value{
Classifiers produce an object which is appropriate
for respective predictor. Predictors returns class
predictions for \code{x_new}.
}
\description{
for internal use
}
\keyword{internal}
