% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.booster.R
\name{predict.booster}
\alias{predict.booster}
\alias{predict.discrete_adaboost}
\alias{predict.real_adaboost}
\title{Prediction function for Adaboost framework}
\usage{
\method{predict}{booster}(object, newdata, type = "pred", print_detail = FALSE, ...)

\method{predict}{discrete_adaboost}(object, newdata, type = "pred", print_detail = FALSE, ...)

\method{predict}{real_adaboost}(object, newdata, type = "pred", print_detail = FALSE, ...)
}
\arguments{
\item{object}{booster object}

\item{newdata}{a factor class variable. Boosting algorithm allows for}

\item{type}{pre-ready or a custom classifier function.}

\item{print_detail}{prints the prediction process. Default is \code{FALSE}.}

\item{...}{additional arguments.}
}
\value{
A vector of class predictions or a matrix of class probabilities
depending of \code{type}
}
\description{
Makes predictions based on booster function
}
\details{
Type "pred" will give class predictions. "prob" will give probabilities for
each class.
}
\seealso{
[predict()]
}
