% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.w_naive_bayes.R
\name{predict.w_naive_bayes}
\alias{predict.w_naive_bayes}
\alias{predict.w_discrete_naive_bayes}
\alias{predict.w_gaussian_naive_bayes}
\title{Predict Discrete Naive Bayes}
\usage{
\method{predict}{w_naive_bayes}(object, newdata = NULL, type = "prob", ...)

\method{predict}{w_discrete_naive_bayes}(object, newdata, type = "prob", ...)

\method{predict}{w_gaussian_naive_bayes}(object, newdata = NULL, type = "prob", ...)
}
\arguments{
\item{object}{"w_bayes" class object..}

\item{newdata}{new observations which predictions will be made on.}

\item{type}{"pred" or "prob".}

\item{...}{additional arguments.}
}
\value{
A vector of class predictions or a matrix of class probabilities
depending of \code{type}
}
\description{
Function for Naive Bayes algorithm prediction.
}
\details{
Calls \code{predict.w_discrete_naive_bayes} or \code{predict.w_gaussian_naive_bayes}
accordingly

Type "pred" will give class predictions. "prob" will give probabilities for
each class.
}
\seealso{
[predict()], [rbooster::predict.w_discrete_naive_bayes()], [rbooster::predict.w_gaussian_naive_bayes()]
}
