% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.booster.R
\name{print.booster}
\alias{print.booster}
\alias{print.discrete_adaboost}
\alias{print.real_adaboost}
\title{Print booster}
\usage{
\method{print}{booster}(x, ...)

\method{print}{discrete_adaboost}(x, ...)

\method{print}{real_adaboost}(x, ...)
}
\arguments{
\item{x}{booster object}

\item{...}{additional arguments.}
}
\value{
Summary of "booster" object.
}
\description{
Prints a summary of booster model
}
\keyword{internal}
