% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_naive_bayes.R
\name{w_naive_bayes}
\alias{w_naive_bayes}
\alias{w_gaussian_naive_bayes}
\alias{w_discrete_naive_bayes}
\title{Naive Bayes algorithm with case weights}
\usage{
w_naive_bayes(x_train, y_train, w = NULL, discretize = TRUE, breaks = 3)

w_gaussian_naive_bayes(x_train, y_train, w = NULL)

w_discrete_naive_bayes(x_train, y_train, breaks = 3, w = NULL)
}
\arguments{
\item{x_train}{explanatory variables.}

\item{y_train}{a factor class variable.}

\item{w}{a vector of case weights.}

\item{discretize}{If \code{TRUE} numerical variables are discretized and discrete naive bayes is applied,}

\item{breaks}{number of break points for discretization. Ignored if \code{discretize = TRUE}.}
}
\value{
a \code{w_naive_bayes} object with below components.
\item{n_train}{Number of cases in the input dataset.}
\item{p}{Number of explanatory variables.}
\item{x_classes}{A list of datasets, which are \code{x_train} separated
for each class.}
\item{n_classes}{Number of cases for each class in input dataset.}
\item{k_classes}{Number of classes in class variable.}
\item{priors}{Prior probabilities.}
\item{class_names}{Names of classes in class variable.}
\item{means}{Weighted mean estimations for each variable.}
\item{stds}{Weighted standart deviation estimations for each variable.}
\item{categories}{Labels for discretized variables.}
\item{boundaries}{Upper and lower boundaries for discretization.}
\item{ps}{probabilities for each variable categories.}
}
\description{
Function for Naive Bayes algorithm classification with case weights.
}
\details{
\code{w_naive_bayes} calls \code{w_gaussian_naive_bayes} or \code{w_discrete_naive_bayes}.

if \code{discrete = FALSE}, \code{w_gaussian_naive_bayes} is called. It uses Gaussian densities with case weights and allows
multiclass classification.

if \code{discrete = TRUE}, \code{w_discrete_naive_bayes} is called. It uses conditional probabilities for each category with
laplace smoothing and allows multiclass classification.
}
\examples{
asd <- rnorm(100)

}
